Always populate list of Tag objects in describe image response
When describing an image, we populate the response using either the
information of the available AMI or, if the image is still under
construction, the information of the CloudFormaton stack. In the case
the image is already available, we populate the response object with a
list of Tag objects (the imageBuilder image's tags). With this patch, we
also return a list of Tag objects (the tags of the imagebuilder's
CloudFormation stack) instead of directly returning a list of key-value
dictionaries as before.