Scope PassRole permissions based on specified instance roles
This commit modifies the PassRole permissions that are attached to the
default head node instance profile. Previously it was assumed that the
head node would need only enable PassRole permissions for IAM roles with
a name like "/parallelcluster/". This default permission is still
included when a custom instance role is not specified.

Signed-off-by: Tim Lane <tilne@amazon.com>