Handle differences between AWS Batch and Slurm in API tests
When Slurm clusters are created, the instances are ready as soon as the
stack creation is completed. With AWS Batch this is not true, so we add
a step to wait until the desired number of instances are present in the
cluster. Furthermore, the queue name for AWS Batch clusters is not
populated (not that we can specify only one queue), and so this patch
handles also that corner case when creating the compute_node_map.