Use "reason" to communicate reason for custom resource status.
Before, we used "response_data" to communicate the custom resource
status. This field, however, is used to set json with key-value pairs
that constitute the set of attributes we could potentially ask the
custom resource for with a !GetAtt statement. In our case, we are not
interested in setting a "Message" attribute or return value, but to
set the reason/status message in the CFN deployment (i.e. in case of
failure, for example, we want to populate that part of the table in
order to disply the failure reason).