Set a physical resource id for the custom resource
By default, the log stream name is used as the custom resource's
physical resource id, which means that it's highly likely that it
would chance across subsequent deployments, leading to a replacement
of the custom resource during an update. In order to avoid that, we
create a physical resource id for the resource.