Add infrastructure to update local ParallelCluster Docker image
Here, by API update we mean the update of the ParallelCluster Docker
image *for a given ParallelCluster version* to the latest build. For
upgrades to a new version of the API we simply re-run the SAM
template. This patch crease an ImageBuilder pipeline, an SNS topic to
use to be notified when the updated image is available, and a Lambda
to make the ParallelClusetr Lambda point to the new Docker image. We
also modify the custom resource responsible for cleanup by making it
also remove all ECR Docker images of the version of the API that we
are deleting or upgrading, and also all the ImageBuilder images
created by the ImageBuilder pipeline (the first one, created by the
CFN template, is delete automatically when updating or deleting the stack).