Upgrade OSU benchmark to 5.7.1 and add other collective and pt2pt benchmarks
OSU pt2pt benchmarks cannot be executed with more than 2 nodes.
We're running it in 2 instances with 1 proc per instance, defined by map-by parameter.

OSU collective benchmarks can be executed with any number of instances,
4 instances are enough to see performance differences.

For all the tests I collected benchmarks results for both Pcluster 2.11.0 and 2.11.1
and using as expected value one close to the 2.11.1 results.

Check avg latency for all packet sizes, with a tolerance of 10%.
Accept a max number of 3 failures on a total of 24 tests.
This helps to avoid failures for minor spikes.

Signed-off-by: Enrico Usai <usai@amazon.com>