Fix cluster config for slurm cluster update test
The config for the slurm cluster update test had EFA enabled for a
compute resource within a queue that did not explicitly enable or
disable the use of a placement group. Starting with 02cc7c14 this
results in a cluster config validation error. cc4cbfed changed some of
the cluster configurations for integration tests affected by this, but
missed the ones modified by this commit.

Signed-off-by: Tim Lane <tilne@amazon.com>