Check bucket existence before using it in exception messages
In the modified methods we were accessing the bucket.name when raising the exception
but if the failure causing the exception is due to the backet absence the raising will fail too.

With this patch we're adding a check for bucket existence out of the global try/catch
using the bucket name in the message.

We can access to the bucket name only in the S3Bucket class but not in the
cluster and imagebuilder module because we cannot be sure the bucket exists.

Signed-off-by: Enrico Usai <usai@amazon.com>