Set cloud-init datasource list
This is already the default for Alinux platform, but not for Ubuntu and Centos.

Adding a datasource directive can save time when cloud-init runs, because prevent
cloud-init to attempt to identify the cloud platform using the script ds-identify and to cycle through all the supported datasources. This check takes around 10 seconds on a c5.2xlarge

Source https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/8/html/configuring_and_managing_cloud-init_for_rhel_8/introduction-to-cloud-init_cloud-content#cloud-init-identifies-the-cloud-platform_introduction-to-cloud-init

Signed-off-by: Luca Carrogu <carrogu@amazon.com>