Inject headnode private ip into compute launch template
Inject headnode private ip into compute launch template, storing it into dna.json used as input for the chef run, so that it's available in the early stages of the cookbook run.
This allows compute to be able to mount the shared folder from the head.

Headnode private ip is got from ENI attribute

Signed-off-by: Luca Carrogu <carrogu@amazon.com>