Disable EnhancedImageMetadataEnabled when building custom image
This option doesn't give advantage for PCluster use case. It is used
to collect metadata of the building instance and these metadata are queryable
against the ImageBuilder Image through the ListImagePackages API. But since the
ImageBuilder Image resource is being deleted after a PCluster successful build,
it doesn't make sense to keep this option enabled. Eventually, we can keep an
option to enable back the enhancedImageMetadata, but this can be placed under a
DevSettings section.

Signed-off-by: Francesco De Martino <fdm@amazon.com>