integ-tests: Fix ad_integration tests
The test was not able to reuse the directory service stack because the random passwords were generated/read in the wrong scope. This commit will generate the passwords in the same scope as the directory service stack. Following code will read the passwords from stack output.

This commit will move the clusters_factory fixture to the last argument of the test to ensure logs are cleaned up after a successful test run.

Signed-off-by: Hanwen <hanwenli@amazon.com>