Disable packages update at instance launch time
The automatic boot update performed by cloud-init can create misalignment in the new compute node nodes and increases boot time.
In this way we're sure that the version of all the packages is aligned in all the instances for the entire cluster lifecycle.

Actually the updates were enabled for AmazonLinux2 only, in this way we're aligning the behaviour of all OSes.

Verified by checking the generated `/var/lib/cloud/instance/cloud-config.txt` file,
verified `/var/log/yum.log` and `/var/log/apt/` are empty,
verified there are no `yum` actions in the `cloud-init-output.log` file
and tested creating a new cluster with custom templates for all the OSes.

Signed-off-by: Enrico Usai <usai@amazon.com>