integ-tests: improve ad integration tests to test LDAPS with certificate verification
LDAPS connection has been tested, but without certificate verification. This commit create a self-signed certificate and instruct sssd agent to trust the certificate.

The reason for changing the `store_secret_in_secret_manager` from troposphere CloudFormation to boto3 calls is that CloudFormation only stores SecretString, but not SecretBytes

This commit also reduce the instance size used for the test.

Signed-off-by: Hanwen <hanwenli@amazon.com>