Improve get_image_log_events test to use the actual first message and not a predefined one
It is not always true that the first message is the cloud init debug message,
sometimes the first message is "STARTED SETUP", causing the test to fail.

A failing test was returning:
`AssertionError: Expected <Setup: STARTED SETUP> to match pattern <Document arn:aws.*:imagebuilder:.*parallelclusterimage.*>, but did not.`

and the message order was:
```
events = [
  {'message': 'Setup: STARTED SETUP', 'timestamp': '2022-01-19T07:24:05.899Z'},
  {'message': 'Setup: Document arn:aws.*:imagebuilder:.*parallelclusterimage.* is an EC2 Image Builder Component Version Arn', 'timestamp': '2022-01-19T07:24:05.900Z'}
]
```

With this patch we're using the actual first message to verify if it's in the list of returned events.
We don't need to check the cloud init debug message is the first one or that it is present.

Signed-off-by: Enrico Usai <usai@amazon.com>