Improve tags management
Created `get_cluster_tags` function to retrieve a list of tags to be added to all the resources
(only PCLUSTER_CLUSTER_NAME_TAG for now).

Refactored code to reuse this function and put in common a `dict_to_cfn_tags` utility function
to convert from a dictionary to a CfnTag list.

Added unit tests for `dict_to_cfn_tags`, `get_cluster_tags` and `get_default_volume_tags` to be sure the
introduced functions and the refactoring are working as expected.

Signed-off-by: Enrico Usai <usai@amazon.com>