Improve get_cluster_log_events test looking for the actual first message
It is not always true that the first message is the cloud init debug message,
sometimes the first message is another one causing the test to fail.

With this patch we're using the actual first message to verify if it's in the list of returned events.
We don't need to check the cloud init debug message is the first one or that it is present.

Signed-off-by: Enrico Usai <usai@amazon.com>