Add head node dependency to cleanup R53 resource
For the head node, add a dependency to the cleanup Route53 resource, so that Route53 Hosted Zone is cleaned after node is deleted.
This will solve random stack delete failed issue due to Route53 hosted zone not empty, occuring when head node is pushing entries in the hosted zone right after the cleanup function is already deleted.

Template diff
* before
```
  HeadNode:
    DependsOn:
    - ComputeFleetLaunchTemplate173566ba834494dcC0292928
    - ComputeFleetTerminateComputeFleetCustomResource9CE1795B
```

* after
```
  HeadNode:
    DependsOn:
    - CleanupRoute53CustomResource
    - ComputeFleetLaunchTemplate173566ba834494dcC0292928
    - ComputeFleetTerminateComputeFleetCustomResource9CE1795B
```

Signed-off-by: Luca Carrogu <carrogu@amazon.com>