Integ-tests: Fix scaling test framework
The thought was to always read `benchmarks` information even when it is None. However, the test complaint of unused parametrization when the `benchmarks` is not used.

Therefore, we only add the `benchmarks` to the parametrization only if the test definition file contains benchmarks specifications. Moreover, instead of consuming the `benchmarks` parameter from `pcluster_config_reader`, which is used by lots of tests without scaling tests, we now consume the `benchmarks` only when the test needs to run scaling tests.

Signed-off-by: Hanwen <hanwenli@amazon.com>