Increase timeout of head node instance to 1h30M
This is required when using Slurm as scheduler and runtime-bake cluster creation,
because the head node waits for the compute node initialization,
but the compute node starts only when the head node is ready.
For ubuntu18 runtime-bake initializiation of the head node takes 23 minutes.

With SGE and Torque there was the ASG managing the `initial_queue_size`
so the compute node started together with the head node.

Note: Alinux2 runtime bake test is passing, this confirms that the startup time depends by the base AMI too.
I tested Centos7, Ubuntu18 and Ubuntu20 and all of them are failing even with a timeout set to 1h.

Signed-off-by: Enrico Usai <usai@amazon.com>