Fix injection of scheduler-definition setting
Avoid to override the scheduler-definition setting into the pytest request, since the same request can be shared by same test during different runs, e.g. when a test is retried because a failure.
It can happen that in between the different runs, the shared fixture containing the scheduler-definition mapping is deleted and then recreated, hence, the request must contain the original scheduler-definition and not the override.

The original scheduler-definition is preserved into the request, and the scheduler-definition url is injected into the scheduler configuration when it is created.

Tested with
```
{%- import 'common.jinja2' as common with context -%}
---
test-suites:
  schedulers:
    test_slurm.py::test_slurm:
      dimensions:
        - regions: ["eu-west-1"]
          instances: {{ common.INSTANCES_DEFAULT_X86 }}
          oss: ["alinux2"]
          schedulers: ["slurm", "my_plugin"]
scheduler-plugins:
  my_plugin:
    scheduler-definition: "../../scheduler_plugins/slurm/utils/upload_artifacts.sh"
    scheduler-commands: "tests.common.schedulers_common.SlurmCommands"
    requires-sudo: true
  my_template:
    scheduler-definition: "../../scheduler_plugins/plugin_template/utils/upload_artifacts.sh"
    scheduler-commands: "tests.common.schedulers_common.SlurmCommands"
    requires-sudo: true
  my_slurm:
    scheduler-definition: "s3://carrogu-bucket-eu-west-1/scheduler_plugins/slurm/plugin_definition.yaml"
    scheduler-commands: "tests.common.schedulers_common.SlurmCommands"
    requires-sudo: true
```

Signed-off-by: Luca Carrogu <carrogu@amazon.com>