Enable traversing of config comparison
Currently, the configuration comparison stops as soon as a new parameter is found in the new configuration.
With this patch, the comparison goes further, traversing also the children of the new parameter when this is a section.
This change allows to be able to ignore the update policy of a section and have the updatability depending on the policies of its children.

Example: changing config from
```
Scheduling:
  SlurmQueues:
  [...]
```
to
```
Scheduling:
  SlurmSettings:
    ScaledownIdletime: 12
  SlurmQueues:
```

the output of update-cluster changes from
```
{
  "message": "Update failure",
  "updateValidationErrors": [
    {
      "parameter": "Scheduling.SlurmSettings",
      "requestedValue": {
        "ScaledownIdletime": 12
      },
      "message": "All compute nodes must be stopped. Stop the compute fleet with the pcluster update-compute-fleet command",
      "currentValue": "-"
    }
  ],
  "changeSet": [
    {
      "parameter": "Scheduling.SlurmSettings",
      "requestedValue": {
        "ScaledownIdletime": 12
      },
      "currentValue": "-"
    }
  ]
}
```
to
```
{
  "message": "Update failure",
  "updateValidationErrors": [
    {
      "parameter": "Scheduling.SlurmSettings.ScaledownIdletime",
      "requestedValue": 12,
      "message": "All compute nodes must be stopped. Stop the compute fleet with the pcluster update-compute-fleet command"
    }
  ],
  "changeSet": [
    {
      "parameter": "Scheduling.SlurmSettings.ScaledownIdletime",
      "requestedValue": 12
    }
  ]
}
```
which is consistent to the output showed when adding a single parameter under a section that already exist

Signed-off-by: Luca Carrogu <carrogu@amazon.com>