Integ-tests: Add benchmark configuration to multiple EFS and FSx
I didn't realize if `run_benchmarks` fixture was used for the tests, the `benchmarks` section has to be defined in the test definition file. Otherwise, the tests would fail for "fixture not exists" error even if the `--benchmarks` flag is not specified with the test_runner. For sure, the benchmark testing framework should be improved. But now, it is easier to just add the `benchmarks` to the test definition

Signed-off-by: Hanwen <hanwenli@amazon.com>