Add change set file into S3 bucket during cluster update
Upload a json file containing the change set to be performed during an update.
The file is in the form
```
{
  "changeSet": [
    {
      "parameter": "Scheduling.SlurmQueues[compute2].ComputeResources[compute2-i1].MinCount",
      "requestedValue": 0,
      "currentValue": 1,
      "updatePolicy": "COMPUTE_FLEET_STOP"
    }
  ]
}
```

The important part is that it will bring information about the update policies of all the changed parameters.
This information is then consumed during the update by the update replacement logic.

Signed-off-by: Luca Carrogu <carrogu@amazon.com>