Do not require Enabled set to true when passing an existing Placement Group Id
If PlacementGroup/Id is configured it is no longer required to set PlacementGroup/Enabled to True.

The alternative was to add a validator to require that both Enabled and Id were set when asking for an existing PG
but we decided to simplify this logic and so providing an Id implies that PG is enabled.

Signed-off-by: Enrico Usai <usai@amazon.com>