integ-tests: remove assert_no_error_in_log from test_efa
Remove the check assert_no_error_in_log in test_efa because sometimes there are insufficient capacity isse during launching instances, but instances are eventually able to launch to run the job. In this case, we don't want to fail the test because of temporary insufficient capacity issue.

Also, this is not the test where we want to validate our scaling or node management logic, test_no_error_in_log is already in slurm test. The test_efa will failed if it can not scale even without the log assertion

Signed-off-by: chenwany <chenwany@amazon.com>