Specify role name in cfn-init call
This allows to be able to create HeadNode outside Cloudformation scope, because by default, if no credentials are specified, CloudFormation checks for stack membership and limits the scope of the cfn-init call to the stack that the instance belongs to.

This is useful in case HeadNode instance created by Cloudformation is terminated, to be able to create a new one using the HeadNode LaunchTemplate.

Signed-off-by: Luca Carrogu <carrogu@amazon.com>