Support mounting existing FSx for Ontap and OpenZFS
1. Add section of FsxOntap and FsxOpenZfs settings in configuration file:
```
SharedStorage:
  - MountDir: /my/mount/point1
    Name: name1
    StorageType: FsxOpenZfs
    FsxOpenZfsSettings:
      VolumeId: fsvol-12345678901234567
  - MountDir: /my/mount/point2
    Name: name2
    StorageType: FsxOntap
    FsxOntapSettings:
      VolumeId: fsvol-12345678901234567
```
2. Use class inheritance in cluster_config.py for FSx for Lustre, Optap, OpenZFS
3. Pass two more attributes to cookbook:
`fsx_volume_junction_paths`: It is a comma separated list. There are strings between the comma if it is FSx for Ontap and OpenZFS. There is empty if it is FSx for Lustre.
`fsx_fs_types`: It is a comma separated list. It tells the cookbook what mount method to use according to the file system (Lustre, Ontap, OpenZFS)
4. Improve FSx security group validator to check different ports for different file system
5. Improve integration test to test the new file system. Move fsx_factory to conftest, so it can be reused by other tests

Test:
Manual tests are done with test_ad_integration and test_cluster_in_no_internet_subnet

Signed-off-by: Hanwen <hanwenli@amazon.com>