Integ-tests: Fix test_scheduler_plugin
The test failed because of https://github.com/aws/aws-parallelcluster/commit/fbb995d4221b56b62cb8d7edd5ffa91ee643f002. The test checks the cluster configuration file with default values on the head node is as expected. Because the root volume configuration was added to the default in the previous commit. The expected outputs have to be changed accordingly. Adding the default root volume configuration makes sense, because those values are used to configure the root volume.

test_scheduler_plugin was successful after this commit.

Signed-off-by: Hanwen <hanwenli@amazon.com>