Extend system-analyzer.sh to include /etc/passwd content
This is useful to run it in a running cluster and understand if the UIDs of the users
are not aligned between head node and compute node.

This might happen for example when using the "Custom AMI at runtime" approach for ParallelCluster 2.x,
on which the new compute nodes might be not aligned with the head node.

Signed-off-by: Enrico Usai <usai@amazon.com>