Adds AWS Budgets support to the configuration file (#4136)
Enables the creation of AWS Budgets using the cluster configuration file, under DevSettings.

Creates schema and configuration for the necessary fields when creating a list of AWS Cost Budgets. Generates the CloudFormation template for the list of budgets in the configuration. Some field validation also added, to ensure proper functionality.

Unit tests added for the validation of the budget schema.

Co-authored-by: Daniel Leite <leitedan@amazon.com>