Add integration tests to check budget creation and update (#4183)
Creating a cluster without a budget and adding the bugdget with an update. Creating a cluster with budgets and changing their properties, validating changes. Uses the boto3 api to describe and verify budget update, creation, and deletion.
Also changed the naming of the budgets, to include a short random alphanumeric string at the end, avoiding name duplication, since the names are the primary identifiers for the budget api.

Co-authored-by: Daniel Leite <leitedan@amazon.com>