Add integration tests for Trainium instances
The test configuration is static, with two trn1.32xlarge and one trn1.2xlarge.
We're running allreduce tests in a single instance on trn1.2xlarge
and collective test (leveraging EFA) on the two trn1.32xlarge nodes.

Some code is temporary and must be removed after GA:
* Installation of neuronx packages from aws-parallelcluster-beta bucket
* Setup of private repository
* Setup of private PyPI repository
* InstanceTypeBaseAMICompatibleValidator usage

The neuron-installation.sh script contains also the instructions to install official neuronx repo
and install dkms driver but they are useless because dkms is already installed in the AMI.

Move tests using c6gn.16xlarge from us-west-2 to eu-west-1 because they are not avail
in the avail zone required for trainium tests.

Change c4.xlarge to c5.large because it is not available in the avail zone required
for traininum tests.

Signed-off-by: Enrico Usai <usai@amazon.com>