Add unit test to verify CFN limits of generated template
A new autogenerated configuration file will be used to try to build a CFN template
and assert that size of the template doesn't exceed 1MB and number of resources doesn't exceed 500.
https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/cloudformation-limits.html

In the new `slurm.full.all_resources.yaml"` configuration file we're defining a number of queues,
compute resources, shared storages up to the limits imposed by the pcluster validators.

Important note: this new config file is very close to the existing limit of 500 AWS resources
and I had to limit the number of Compute Resources for half of the queues to not exceed CFN
template size limits.

Signed-off-by: Enrico Usai <usai@amazon.com>