Add dryrun validator test SlurmFlexibleComputeResource
Add dryrun validator test SlurmFlexibleComputeResource to test only the first InstanceType in the list, through a RunInstances.

Calling a dryrun for CreateFleet is not possible at this point, because the LaunchTemplate doesn't exist yet.

The purpose of the test is to check if it's possible to launch an instance with all the other configurations in place, e.g. AMI ID, security groups, subnet, etc.

Signed-off-by: Luca Carrogu <carrogu@amazon.com>