Add validation recipe call to image validation (#4372)
Added call to aws-parallelcluster::validation recipe to the BuilderImage
validation component template in order to allow validation to be
carried out within chef. This should enable us to move validation
to chef recipes with easier access to node configuration. This should
help us not exceed the maximum number of steps allowed in a component
because we can now add validation to the chef cookbooks instead of
adding new steps to the template.

Added validation recipe for validation MySQL client package installation.