Create an update policy for managed placement groups
Managed placement group deletion requires the compute fleet to be stopped.
The update policy for placement groups and all of the parameters within
needs to check for deletion when updating the cluster.  Also, the networking
section of the compute resource only contains placement group, so that section
follows this policy as well.

Signed-off-by: Ryan Anderson <ndry@amazon.com>