Create an update policy for managed placement groups (#4393)
Managed placement group deletion requires the compute fleet to be stopped.
The update policy for placement groups and all of the parameters within
needs to check for deletion when updating the cluster.  Also, the networking
section of the compute resource only contains placement group, so that section
follows this policy as well.

Changes at the Queue level will be treated as global and require the fleet to be stopped.  Changes
at the Compute Resource level will be evaluated against the queue to ensure actual deletion.

Signed-off-by: Ryan Anderson <ndry@amazon.com>