Add integration tests for Slurm Accounting
This change adds integration tests to make sure
that, if configured, Slurm accounting is properly
configured, running, and recording Slurm jobs.

It also checks the a cluster can be updated from
not having Slurm accounting configured to having
it configured.

As part of the tests, get_log_streams was updated
to support more than one page of log streams.

Signed-off-by: David Pratt <davprat@amazon.com>