Require IMDSv2 in PC API Cfn infrastructure
This patch requires IMDSv2 for the ImageBuilder build instance
used when setting up the PC API infrastructure with CloudFormation,
which we use to copy the Docker image used by the Lambda into the
customer's private ECR repository. This change is safe because
there are no custom components that do `curl` calls with IMDSv2,
(see the `EcrImageRecipe` resource).