Change method returning placement group settings
Before the method was called `get_placement_group_key_for_compute_resource` and it was returning a tuple, not just the key, so it was not clear from the caller the returned output (without knowing in advance was the method was returing.

Method is now called get_placement_group_settings_for_compute_resource and it's returning a dict.

Signed-off-by: Luca Carrogu <carrogu@amazon.com>