Catch AWSClientError when caching capacity reservation information
Since this cache is only used for validation, if AWSClientError happens (e.g insufficient IAM permissions to describe the capacity reservations), we catch the exception to avoid blocking CLI execution if the user want to suppress the validation. In other words, we delay any boto3 exceptions from caching phase to validator execution phase.

Signed-off-by: Hanwen <hanwenli@amazon.com>