Add missing fields to launch template validators
This patch adds the missing fields to the launch template validators.
For the head node, these are the `BlockDeviceMappings` and `KeyName`,
while for the compute resources these are the `BlockDeviceMappings`,
`InstanceMetadataOptions` and `CapacityReservationSpecification`.

This patch extracts the common code into functions that generate
dictionaries (or lists of dictionaries), which are consumed directly by
the validators (they do a run instance dry run, as they cannot validate
the launch templates produced via CDK), and converted into CDK objects
in the case of the code that actually generate the launch templates via
CDK/CloudFormation.

Signed-off-by: Ermanno Moser <ermann@amazon.com>