Change Node version validator to use 'nvm current'
If the call to nvm fails, the validator falls back to calling
'node --version'. If the node major version is detected to be
greater than a max major version, a warning is logged directing
the user to install and use the latest version of the max node
version if they have problems with the current node version.

Signed-off-by: David Pratt <davprat@amazon.com>