Remove creation of EFS mount targets for existing FS
Before this patch PCluster create EFS mount targets for existing FS if there isn't one already in the AZs of the cluster. Such mount target resource is therefore associated with a cluster lifecycle and is deleted when the cluster is deleted.

Due to the fact that each EFS can have a single mount target per AZ, in case the user mounts the same FS in multiple clusters, when the first one is deleted all other clusters will see the mount target revoked and will break.

Following the same approach for all un-managed PCluster resources, within this patch it is removed the creation of missing mount target for existing EFS and simply fails at validation time when requirements are not met.

Signed-off-by: Luca Carrogu <carrogu@amazon.com>