Remove empty capacity reservation id/resource group ARN from dict in launch template validator
Capacity reservations specifications can have either a capacity reservation ID or a
capacity reservation resource group ARN, but not both at the same time. This patch
removes the one of the two which has None value from the dictionary of the capacity
reservation that is passed to the launch template validator's RunInstances DryRun, as
the RunInstances DryRun breaks when it is passed a None as an argument (since it expects
a String). Note that by contrast the actual CDK capacity reservation specification construct
that we use to create the actual launch template accepts None as an argument.