Remove constraint of 1 subnet ID for Slurm and Scheduler Plugin queues
Move the definition of the subnet_ids field to the children classes
SlurmQueueNetworkingSchema and AwsBatchQueueNetworkingSchema.
In case of AWS Batch queues the constraint remains the same.
In case of Slurm and Scheduler Plugin queues, the new constraint is
that at least one subnet must be defined.

Split schema unit tests for head node, AWS Batch queues and Slurm
and Scheduler Plugin queues.

Signed-off-by: Jacopo De Amicis <jdamicis@amazon.it>