fix easyconfig in an ipv6-only subnet environment
previously the configuration generator would crash if it encountered an
IPv6-only subnet in an AWS VPC.

This would take place during the VPC/subnet probe process and simply
take place if it was in the same AWS account (Before you can select a
VPC to use for pcluster).

This takes place in _extract_subnet_size() as 'CidrBlock' is only
present for IPv4-enable subnets.

This adds a simple filter to ignore IPv6-only subnets for now (ipv6-native is
false).

This will *not* impact dual-stack subnets (IPv4 and IPv6).