Fix validator to avoid managed FSx storage
Fix validator to avoid managed FSx storage with different single subnets in different queues.
Validator check if managed FSx are set in cluster configuration and fail if there is at least one set. The case is when a single subnet is set for a queue, but each queue use a different subnet.

FSx for Lustre cannot be created passing a list of subnets, even if all the subnet are in the same AZ.

Signed-off-by: Luca Carrogu <carrogu@amazon.com>