Make updates fail in case of failures during cfn-hup actions
Previously, if during a cluster update errors occurred during cfn-hup
actions, such as a failure while running the cookbooks, in
CloudFormation the update completed successfully. This patch alters
the hook that is called during the cfn-hup so that in case errors
occur during cfn-hup actions the CloudFormation update fails with an
error message.