Print changes in the vpc_security_group_id when the update is allowed
Without this patch the update command wasn't displaying any change in the config
when the update of the vpc_security_group_id parameter was allowed (not managed fsx).

Invert test logic to verify all the expected changes are in the output and not viceversa.
This logic was hiding the fact we were not printing the change in case of allowed update
and was hiding a typo in the check result.

Signed-off-by: Enrico Usai <usai@amazon.com>