Fix test_on_demand_capacity_reservation
There are two kinds of failure of this test:
1. The fixtures to create capacity reservations fail because of insufficient capacity.
Solution: Change the instance types to more common ones. Note that burstable instance types do not support placement groups. Therefore, I have to use c5, r5, m5 instances.
2. The test started to fail constantly after https://github.com/aws/aws-parallelcluster/pull/4484. Because the PR added CapacityReservationSpecification in the CLI ec2 run_instances dryrun validator. The PR itself is correct, but uncovered an issue about placement group. With ODCR support, we started supporting placement groups on both queue and compute resource level. But the dryrun still used placement group only on queue level. As a result, the dryrun validation would fail if the capacity reservation in CapacityReservationSpecification required a placement group and the placement group was specified on compute resource level.
Solution: Use the right placement group in the dry run. To achieve this, the dry run validation needs to be in an inner loop

Signed-off-by: Hanwen <hanwenli@amazon.com>