Split p4d EFA and multi-NICs tests in separate files
EFA tests don't need p4d in the head node and they can be executed with the current code.

Instead, to run multi-NICs test it is required to force usage of ODCR for the head node
by adding `CapacityReservationSpecification` in the launch template.

By splitting the tests is easier to execute them separately.

Signed-off-by: Enrico Usai <usai@amazon.com>