Make pcluster configure test prompts compatible with instances with EFA
If the instance has EFA support (like hpc6id) the prompt will ask to enable EFA
and if answering in a positive way, will ask for placement group too.

In this way the code is generic enough to support both the cases.
I tested it using hpc6id and c5.xlarge as instances.

Signed-off-by: Enrico Usai <usai@amazon.com>