Limit SSH connection access from instance running the test
By default the `AllowedIps` value is set to 0.0.0.0/0
with this patch we're limiting the access to the head node by default from instance running the test
if it is an EC2 instance.

This SSH access is used by using the SSH key to execute actions in the head node (e.g. job submission)
and by using the SSH user/password for example by AD tests to verify AD setup is working as expected.

Signed-off-by: Enrico Usai <usai@amazon.com>