[MultiAz] Ensure subnets in VPC stack are in different AZ when at least two are available
Subnets in vpc_stacks are assigned to AZ that are picked from a list.
When there are overrides the limit the number of available AZ, the element in the list
may be duplicated.
This change ensure that the first two element of the list are always different if at least
two AZ are available in the region.

Signed-off-by: Nicola Sirena <nssirena@amazon.it>