Improve test for dynamic nodes not powering down after cluster update
Improve test for dynamic nodes not powering down after cluster update:
the new test manually trigger a slurmctld restart and scontrol reconfigure until they are executed in the same timestamp second, so to trigger the issue that was present up to Slurm < 22.05.6

Executed the test with PCluster 3.2.1, which ships Slurm 21.08.8' and got a failure
```
>           assert_that(expected_states).contains(node_states.get(node))
E           AssertionError: Expected <['idle~']> to contain item <idle>, but did not.
```

Executed the test with Slurm 22.05.7 and got no failure

ref. https://bugs.schedmd.com/show_bug.cgi?id=13953

Signed-off-by: Luca Carrogu <carrogu@amazon.com>