Validate cluster name length with Slurm accounting
Add condition to the ClusterNameValidator to check that the cluster
name is not longer than 40 chars.

When enabling Slurm accounting, Slurm creates database tables with a
name equal to "cluster_name" + "suffix". MySQL and MariaDB have a
maximum table name length of 64 chars. The maximum suffix length used
by Slurm is 24 chars. Therefore, when Slurm accounting is enabled the
cluster name must not be longer than 40 chars.

Add new unit test for ClusterNameValidator.

Signed-off-by: Jacopo De Amicis <jdamicis@amazon.it>