Add integration test for scontrol update nodelist sorting bug
Slurm 22.05 (up to 22.05.7) sorts the nodes in nodelist provided
as nodename field to the `scontrol update` command.
If `scontrol update nodename=nodelist nodeaddr=nodeaddrlist` is given,
this causes a mismatch between the nodenames and the nodeaddrs
because the order of nodeaddrlist is not changed to match the reordering
of nodelist.

The bug above will be fixed in Slurm 22.05.8 and later.

The added test checks that the sorting of the nodelist does not happen.

Signed-off-by: Jacopo De Amicis <jdamicis@amazon.it>