update the general cfn-init error message (#4801)
When the head node fails bootstrap, we would like to show some general "catch-all" error message if no specific error messages are available. The previous message "Failed to run cfn-init ......" is not customer-friendly since it requires understanding of internal mechanism (cfn-init). We update the error message here in this PR.