`[lib]` Add a library for accessing parallelcluster.
This commit adds a library for accessing parallelcluster by generating the
class from the spec. It does more type checking as the CLI is able to do
that portion through the argument parsing library, as well as accepting
Python datastructures for those parameters of type "file" such as cluster
and image configuration.