Use custom cookbook to solve the Ubuntu20 build failure due to MySQL installation
Build of Ubuntu 20.04 x86_64 custom images started to fail on 25 January,
because an updated version of MySQL packages (8.0.32) has been released in the Ubuntu OS repository.

This package is installed in the system by the initial phases of the build AMI process
and then the install_mysql_client recipe code tries to install the old 8.0.31 version
on top of the updated version, causing the installation steps to fail.

This patched cookbook solves the issue and it is the suggested workaround for our customers.

Signed-off-by: Enrico Usai <usai@amazon.com>